requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
discord==2.1.0
discord.py==2.1.0
frozenlist==1.3.3
html5lib==1.1
idna==3.4
multidict==6.0.2
numpy==1.23.5
pandas==1.5.2
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
tabulate==0.9.0
urllib3==1.26.13
webencodings==0.5.1
yarl==1.8.1
